update logs for qdrant collection
The logs for managing the default collections didn't reflect the config settings for that collection name. I also noticed that the collection name was read several different times with essentially the same code. I removed that duplication, and read it once in the init() function and used in all the same places.

Now the logs reflect the name of the collection actually used for the current node (I noticed this because I was running multiple nodes on the same machine - so I have to use different collection names for different running models).